+++ YouboraLib/infinity/YBInfinity.m
@@ -16,8 +16,7 @@
 
 @interface YBInfinity()
 
-// Delegates list
-@property (nonatomic, strong) NSMutableArray<id<YBInfinityDelegate>> * eventDelegates;
+@property (nonatomic, strong) YBInfinityLocalManager* infinityStorage;
 
 @end
 
@@ -52,15 +51,11 @@
 - (void) fireSessionStartWithScreenName: (NSString *) screenName andDimensions:(NSDictionary<NSString *, NSString *> *) dimensions{
     [self generateNewContext];
     
-    for (id<YBInfinityDelegate> delegate in self.eventDelegates) {
-        [delegate youboraInfinityEventSessionStartWithScreenName:screenName andDimensions:dimensions];
-    }
+    [self.delegate youboraInfinityEventSessionStartWithScreenName:screenName andDimensions:dimensions];
 }
 
 - (void) fireNavWithScreenName: (NSString *) screenName {
-    for (id<YBInfinityDelegate> delegate in self.eventDelegates) {
-        [delegate youboraInfinityEventNavWithScreenName:screenName];
-    }
+    [self.delegate youboraInfinityEventNavWithScreenName:screenName];
 }
 
 - (void) fireEvent: (NSDictionary<NSString *, NSString *> *) dimensions values: (NSDictionary<NSString *, NSNumber *> *) values andEventName: (NSString *) eventName {
@@ -80,15 +75,15 @@
         eventName = @"Unknown";
     }
     
-    for (id<YBInfinityDelegate> delegate in self.eventDelegates) {
-        [delegate youboraInfinityEventEventWithDimensions:dimensions values:values andEventName:eventName];
+    if (self.delegate) {
+        [self.delegate youboraInfinityEventEventWithDimensions:dimensions values:values andEventName:eventName];
     }
 }
 
 - (void) fireSessionStop: (NSDictionary<NSString *, NSString *> *) params {
     [self.flags reset];
-    for (id<YBInfinityDelegate> delegate in self.eventDelegates) {
-        [delegate youboraInfinityEventSessionStop:params];
+    if (self.delegate) {
+        [self.delegate youboraInfinityEventSessionStop:params];
     }
 }
 
@@ -109,36 +104,6 @@
 
 - (NSNumber *) getLastSent {
     return [YBInfinityLocalManager getLastActive];
-}
-
-/*- (void) addActiveSession: (nullable NSString *) sessionId {
-    if (sessionId == nil)
-        return;
-    if (self.activeSessions == nil)
-        self.activeSessions = [[NSMutableArray alloc] initWithCapacity:1];
-    [self.activeSessions addObject:sessionId];
-}
-
-- (void) removeActiveSession: (nullable NSString *) sessionId {
-    if (sessionId == nil)
-        return;
-    [self.activeSessions removeObject:sessionId];
-}*/
-
-- (void)addYouboraInfinityDelegate:(id<YBInfinityDelegate>)delegate {
-    if (delegate != nil) {
-        if (self.eventDelegates == nil) {
-            self.eventDelegates = [NSMutableArray arrayWithObject:delegate];
-        } else if (![self.eventDelegates containsObject:delegate]) {
-            [self.eventDelegates addObject:delegate];
-        }
-    }
-}
-
-- (void)removeYouboraInfinityDelegate:(id<YBInfinityDelegate>)delegate {
-    if (delegate != nil && self.eventDelegates != nil) {
-        [self.eventDelegates removeObject:delegate];
-    }
 }
 
 -(NSString* _Nonnull) getSessionRoot {