+++ YouboraLib/transform/resourceparser/YBLocationParser.swift
@@ -31,4 +31,8 @@ import Foundation
         
         return response.allHeaderFields["Location"] as? String
     }
+    
+    public func parseTransportFormat(data: Data?, response: HTTPURLResponse?, listenerParents: [String: AnyObject]?, userDefinedTransportFormat: String?) -> String? {
+        return nil
+    }
 }